requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
pytest>=7.0.0
pytest-cov>=4.0.0
RestrictedPython>=7.0
